requirements.txt
langchain
# openai
Pillow
# stability-sdk
streamlit
streamlit-extras
webuiapi